requirements.txt
streamlit==1.35.0
scikit-learn==1.4.1
pandas==2.2.2
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
